update text spotting into coverage
