+++ .github/workflows/mr_ci.yml
@@ -42,7 +42,7 @@ jobs:
 
       - name: Test with pytest
         run: |
-          python -m pytest --cov=./ --cov-report=xml ./tests
+          python -m pytest ./tests
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
