+++ tests/test_text_spotting/test_dptext_runner.py
@@ -7,24 +7,24 @@ import pickle
 import geopandas as gpd
 import pandas as pd
 import pytest
-from deepsolo.config import get_cfg
 from detectron2.engine import DefaultPredictor
 from detectron2.structures.instances import Instances
+from dptext_detr.config import get_cfg
 
-from mapreader import DeepSoloRunner
+from mapreader import DPTextDETRRunner
 from mapreader.load import MapImages
 
-# use cloned DeepSolo path if running in github actions
-DEEPSOLO_PATH = (
-    pathlib.Path("./DeepSolo/").resolve()
+# use cloned DPText-DETR path if running in github actions
+DPTEXT_DETR_PATH = (
+    pathlib.Path("./DPText-DETR/").resolve()
     if os.getenv("GITHUB_ACTIONS") == "true"
-    else pathlib.Path(os.getenv("DEEPSOLO_PATH")).resolve()
+    else pathlib.Path(os.getenv("DPTEXT_DETR_PATH")).resolve()
 )
 
 
 @pytest.fixture
 def sample_dir():
-    return pathlib.Path(__file__).resolve().parent.parent / "tests" / "sample_files"
+    return pathlib.Path(__file__).resolve().parent.parent / "sample_files"
 
 
 @pytest.fixture
@@ -46,7 +46,7 @@ def init_dataframes(sample_dir, tmp_path):
 @pytest.fixture(scope="function")
 def mock_response(monkeypatch, sample_dir):
     def mock_pred(self, *args, **kwargs):
-        with open(f"{sample_dir}/patch-0-0-800-40-deepsolo-pred.pkl", "rb") as f:
+        with open(f"{sample_dir}/patch-0-0-800-40-dptext-detr-pred.pkl", "rb") as f:
             outputs = pickle.load(f)
         return outputs
 
@@ -56,10 +56,10 @@ def mock_response(monkeypatch, sample_dir):
 @pytest.fixture
 def init_runner(init_dataframes):
     parent_df, patch_df = init_dataframes
-    runner = DeepSoloRunner(
+    runner = DPTextDETRRunner(
         patch_df,
         parent_df=parent_df,
-        cfg_file=f"{DEEPSOLO_PATH}/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        cfg_file=f"{DPTEXT_DETR_PATH}/configs/DPText_DETR/ArT/R_50_poly.yaml",
     )
     return runner
 
@@ -73,69 +73,70 @@ def runner_run_all(init_runner, mock_response):
 
 def test_get_cfg():
     cfg = get_cfg()
-    assert "TEMPERATURE" in cfg["MODEL"]["TRANSFORMER"].keys()
+    assert "USE_POLYGON" in cfg["MODEL"]["TRANSFORMER"].keys()
+    assert "DET_ONLY" in cfg["TEST"].keys()
 
 
-def test_deepsolo_init(init_dataframes):
+def test_dptext_init(init_dataframes):
     parent_df, patch_df = init_dataframes
-    runner = DeepSoloRunner(
+    runner = DPTextDETRRunner(
         patch_df,
         parent_df=parent_df,
-        cfg_file=f"{DEEPSOLO_PATH}/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        cfg_file=f"{DPTEXT_DETR_PATH}/configs/DPText_DETR/ArT/R_50_poly.yaml",
     )
-    assert isinstance(runner, DeepSoloRunner)
+    assert isinstance(runner, DPTextDETRRunner)
     assert isinstance(runner.predictor, DefaultPredictor)
     assert isinstance(runner.parent_df.iloc[0]["coordinates"], tuple)
     assert isinstance(runner.patch_df.iloc[0]["coordinates"], tuple)
 
 
-def test_deepsolo_init_str(init_dataframes, tmp_path):
+def test_dptext_init_str(init_dataframes, tmp_path):
     parent_df, patch_df = init_dataframes
     parent_df = parent_df.to_csv(f"{tmp_path}/parent_df.csv")
     patch_df = patch_df.to_csv(f"{tmp_path}/patch_df.csv")
-    runner = DeepSoloRunner(
+    runner = DPTextDETRRunner(
         f"{tmp_path}/patch_df.csv",
         parent_df=f"{tmp_path}/parent_df.csv",
-        cfg_file=f"{DEEPSOLO_PATH}/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        cfg_file=f"{DPTEXT_DETR_PATH}/configs/DPText_DETR/ArT/R_50_poly.yaml",
     )
-    assert isinstance(runner, DeepSoloRunner)
+    assert isinstance(runner, DPTextDETRRunner)
     assert isinstance(runner.predictor, DefaultPredictor)
     assert isinstance(runner.parent_df.iloc[0]["coordinates"], tuple)
     assert isinstance(runner.patch_df.iloc[0]["coordinates"], tuple)
 
 
-def test_deepsolo_init_pathlib(init_dataframes, tmp_path):
+def test_dptext_init_pathlib(init_dataframes, tmp_path):
     parent_df, patch_df = init_dataframes
     parent_df = parent_df.to_csv(f"{tmp_path}/parent_df.csv")
     patch_df = patch_df.to_csv(f"{tmp_path}/patch_df.csv")
-    runner = DeepSoloRunner(
+    runner = DPTextDETRRunner(
         pathlib.Path(f"{tmp_path}/patch_df.csv"),
         parent_df=pathlib.Path(f"{tmp_path}/parent_df.csv"),
-        cfg_file=f"{DEEPSOLO_PATH}/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        cfg_file=f"{DPTEXT_DETR_PATH}/configs/DPText_DETR/ArT/R_50_poly.yaml",
     )
-    assert isinstance(runner, DeepSoloRunner)
+    assert isinstance(runner, DPTextDETRRunner)
     assert isinstance(runner.predictor, DefaultPredictor)
     assert isinstance(runner.parent_df.iloc[0]["coordinates"], tuple)
     assert isinstance(runner.patch_df.iloc[0]["coordinates"], tuple)
 
 
-def test_deepsolo_init_tsv(init_dataframes, tmp_path):
+def test_dptext_init_tsv(init_dataframes, tmp_path):
     parent_df, patch_df = init_dataframes
     parent_df = parent_df.to_csv(f"{tmp_path}/parent_df.tsv", sep="\t")
     patch_df = patch_df.to_csv(f"{tmp_path}/patch_df.tsv", sep="\t")
-    runner = DeepSoloRunner(
+    runner = DPTextDETRRunner(
         f"{tmp_path}/patch_df.tsv",
         parent_df=f"{tmp_path}/parent_df.tsv",
         delimiter="\t",
-        cfg_file=f"{DEEPSOLO_PATH}/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        cfg_file=f"{DPTEXT_DETR_PATH}/configs/DPText_DETR/ArT/R_50_poly.yaml",
     )
-    assert isinstance(runner, DeepSoloRunner)
+    assert isinstance(runner, DPTextDETRRunner)
     assert isinstance(runner.predictor, DefaultPredictor)
     assert isinstance(runner.parent_df.iloc[0]["coordinates"], tuple)
     assert isinstance(runner.patch_df.iloc[0]["coordinates"], tuple)
 
 
-def test_deepsolo_run_all(init_runner, mock_response):
+def test_dptext_run_all(init_runner, mock_response):
     runner = init_runner
     # dict
     out = runner.run_all()
@@ -145,11 +146,11 @@ def test_deepsolo_run_all(init_runner, mock_response):
     # dataframe
     out = runner._dict_to_dataframe(runner.patch_predictions, geo=False, parent=False)
     assert isinstance(out, pd.DataFrame)
-    assert set(out.columns) == set(["image_id", "geometry", "text", "score"])
+    assert set(out.columns) == set(["image_id", "geometry", "score"])
     assert "patch-0-0-800-40-#mapreader_text.png#.png" in out["image_id"].values
 
 
-def test_deepsolo_convert_to_parent(runner_run_all, mock_response):
+def test_dptext_convert_to_parent(runner_run_all, mock_response):
     runner = runner_run_all
     # dict
     out = runner.convert_to_parent_pixel_bounds()
@@ -159,13 +160,11 @@ def test_deepsolo_convert_to_parent(runner_run_all, mock_response):
     # dataframe
     out = runner._dict_to_dataframe(runner.parent_predictions, geo=False, parent=True)
     assert isinstance(out, pd.DataFrame)
-    assert set(out.columns) == set(
-        ["image_id", "patch_id", "geometry", "text", "score"]
-    )
+    assert set(out.columns) == set(["image_id", "patch_id", "geometry", "score"])
     assert "mapreader_text.png" in out["image_id"].values
 
 
-def test_deepsolo_convert_to_parent_coords(runner_run_all, mock_response):
+def test_dptext_convert_to_parent_coords(runner_run_all, mock_response):
     runner = runner_run_all
     # dict
     out = runner.convert_to_coords()
@@ -175,23 +174,21 @@ def test_deepsolo_convert_to_parent_coords(runner_run_all, mock_response):
     # dataframe
     out = runner._dict_to_dataframe(runner.geo_predictions, geo=True, parent=True)
     assert isinstance(out, gpd.GeoDataFrame)
-    assert set(out.columns) == set(
-        ["image_id", "patch_id", "geometry", "crs", "text", "score"]
-    )
+    assert set(out.columns) == set(["image_id", "patch_id", "geometry", "crs", "score"])
     assert "mapreader_text.png" in out["image_id"].values
     assert out.crs == runner.parent_df.crs
 
 
-def test_deepsolo_deduplicate(sample_dir, tmp_path, mock_response):
+def test_dptext_deduplicate(sample_dir, tmp_path, mock_response):
     maps = MapImages(f"{sample_dir}/mapreader_text.png")
     maps.add_metadata(f"{sample_dir}/mapreader_text_metadata.csv")
     maps.patchify_all(patch_size=800, path_save=tmp_path, overlap=0.5)
     maps.check_georeferencing()
     parent_df, patch_df = maps.convert_images()
-    runner = DeepSoloRunner(
+    runner = DPTextDETRRunner(
         patch_df,
         parent_df=parent_df,
-        cfg_file=f"{DEEPSOLO_PATH}/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        cfg_file=f"{DPTEXT_DETR_PATH}/configs/DPText_DETR/ArT/R_50_poly.yaml",
     )
     _ = runner.run_all()
     out = runner.convert_to_parent_pixel_bounds(deduplicate=False)
@@ -209,7 +206,7 @@ def test_deepsolo_deduplicate(sample_dir, tmp_path, mock_response):
     assert len_07 >= len_05
 
 
-def test_deepsolo_run_on_image(init_runner, mock_response):
+def test_dptext_run_on_image(init_runner, mock_response):
     runner = init_runner
     out = runner.run_on_image(
         runner.patch_df.iloc[0]["image_path"], return_outputs=True
@@ -219,57 +216,11 @@ def test_deepsolo_run_on_image(init_runner, mock_response):
     assert isinstance(out["instances"], Instances)
 
 
-def test_deepsolo_save_to_geojson(runner_run_all, tmp_path, mock_response):
+def test_dptext_save_to_geojson(runner_run_all, tmp_path, mock_response):
     runner = runner_run_all
     _ = runner.convert_to_coords()
     runner.save_to_geojson(f"{tmp_path}/text.geojson")
     assert os.path.exists(f"{tmp_path}/text.geojson")
     gdf = gpd.read_file(f"{tmp_path}/text.geojson")
     assert isinstance(gdf, gpd.GeoDataFrame)
-    assert set(gdf.columns) == set(
-        ["image_id", "patch_id", "geometry", "crs", "text", "score"]
-    )
-
-
-def test_deepsolo_search_preds(runner_run_all, mock_response):
-    runner = runner_run_all
-    _ = runner.convert_to_parent_pixel_bounds()
-    out = runner.search_preds("map", ignore_case=True)
-    assert isinstance(out, dict)
-    assert "mapreader_text.png" in out.keys()
-    # test dataframe
-    out = runner.search_preds("map", ignore_case=True, return_dataframe=True)
-    assert isinstance(out, pd.DataFrame)
-    assert set(out.columns) == set(
-        ["image_id", "patch_id", "geometry", "text", "score"]
-    )
-    assert "mapreader_text.png" in out["image_id"].values
-    out = runner.search_preds("somethingelse", ignore_case=True, return_dataframe=True)
-    assert len(out) == 0
-
-
-def test_deepsolo_search_preds_errors(runner_run_all, mock_response):
-    runner = runner_run_all
-    with pytest.raises(ValueError, match="No parent predictions found"):
-        runner.search_preds("maps", ignore_case=True)
-
-
-def test_deepsolo_save_search_results(runner_run_all, tmp_path, mock_response):
-    runner = runner_run_all
-    _ = runner.convert_to_parent_pixel_bounds()
-    out = runner.search_preds("map", ignore_case=True)
-    assert isinstance(out, dict)
-    runner.save_search_results_to_geojson(f"{tmp_path}/search_results.geojson")
-    assert os.path.exists(f"{tmp_path}/search_results.geojson")
-    gdf = gpd.read_file(f"{tmp_path}/search_results.geojson")
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert set(gdf.columns) == set(
-        ["image_id", "patch_id", "geometry", "crs", "text", "score"]
-    )
-    assert "mapreader_text.png" in gdf["image_id"].values
-
-
-def test_deepsolo_save_search_results_errors(runner_run_all, tmp_path, mock_response):
-    runner = runner_run_all
-    with pytest.raises(ValueError, match="No results to save"):
-        runner.save_search_results_to_geojson(f"{tmp_path}/test.geojson")
+    assert set(gdf.columns) == set(["image_id", "patch_id", "geometry", "crs", "score"])
